+++ src/headless/converse-core.js
@@ -103,7 +103,7 @@ const _converse = {
     'promises': {}
 }
 
-_converse.VERSION_NAME = "v5.0.4dev";
+_converse.VERSION_NAME = "v5.0.4";
 
 Object.assign(_converse, Backbone.Events);