+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Bugfix: Don't treat every duplicate message ID as a message correction; since some
   clients don't use globally unique ID's this causes false positives.
+- #1712: `TypeError: plugin._features is not a function`
 
 ## 5.0.3 (2019-09-13)