+++ CHANGES.md
@@ -8,6 +8,8 @@
 - Bugfix: process stanzas from mam one-by-one in order to correctly process message
   receipts
 - #1714 Bugfix: Don't notify the user in case we're receiving a message delivery receipt only
+- Bugfix: only send bookmark stanza if connected in order to avoid error mesage
+  `improper-addressing` after logging out
 
 ## 5.0.3 (2019-09-13)